Common package refactor done
