+++ resources/lib/common/addon_settings.py
