+++ resources/lib/common/private_func.py
@@ -2,7 +2,33 @@ import socket
 import struct
 
 
-def wake_on_lan(macaddress):
+def is_ip(address):
+    """from http://www.seanelavelle.com/2012/04/16/checking-for-a-valid-ip-in-python/"""
+    try:
+        socket.inet_aton(address)
+        ip = True
+    except socket.error:
+        ip = False
+
+    return ip
+
+
+def wake_servers(settings):
+    if settings.get_setting('wolon'):
+
+        print "PleXBMC -> Wake On LAN: true"
+        for servers in settings.get_wakeservers():
+            if servers:
+                try:
+                    print "PleXBMC -> Waking server with MAC: %s" % servers
+                    __wake_on_lan(servers)
+                except ValueError:
+                    print "PleXBMC -> Incorrect MAC address format for server %s" % servers
+                except:
+                    print "PleXBMC -> Unknown wake on lan error"
+
+
+def __wake_on_lan(macaddress):
     """ Switches on remote computers using WOL. """
 
     macaddress = macaddress.strip()
