+++ resources/lib/plexgdm.py
@@ -20,10 +20,11 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
 MA 02110-1301, USA.
 """
+import socket
+import struct
 
 __author__ = 'DHJ (hippojay) <plex@h-jay.com>'
 
-import struct
 import threading
 import time
 import urllib2
@@ -32,7 +33,6 @@ from common import *
 
 
 class plexgdm:
-
     def __init__(self, debug=0):
 
         self.discover_message = 'M-SEARCH * HTTP/1.0'
@@ -54,10 +54,11 @@ class plexgdm:
         self.discovery_complete = False
         self.client_registered = False
         self.debug = debug
-        self.__printDebug=PrintDebug("PleXBMC", "PlexGDM")
+        self.__printDebug = PrintDebug("PleXBMC", "PlexGDM")
 
     def clientDetails(self, c_id, c_name, c_post, c_product, c_version):
-        self.client_data = "Content-Type: plex/media-player\nResource-Identifier: %s\nName: %s\nPort: %s\nProduct: %s\nVersion: %s" % ( c_id, c_name, c_post, c_product, c_version )
+        self.client_data = "Content-Type: plex/media-player\nResource-Identifier: %s\nName: %s\nPort: %s\nProduct: %s\nVersion: %s" % (
+        c_id, c_name, c_post, c_product, c_version)
         self.client_id = c_id
 
     def getClientDetails(self):
@@ -66,34 +67,36 @@ class plexgdm:
 
         return self.client_data
 
-    def client_update (self):
+    def client_update(self):
         update_sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM, socket.IPPROTO_UDP)
 
-        #Set socket reuse, may not work on all OSs.
+        # Set socket reuse, may not work on all OSs.
         try:
-            update_sock.setsockopt(socket.SOL_SOCKET,socket.SO_REUSEADDR,1)
+            update_sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         except:
             pass
 
-        #Attempt to bind to the socket to recieve and send data.  If we can;t do this, then we cannot send registration
+        # Attempt to bind to the socket to recieve and send data.  If we can;t do this, then we cannot send registration
         try:
-            update_sock.bind(('0.0.0.0',self.client_update_port))
+            update_sock.bind(('0.0.0.0', self.client_update_port))
         except:
-            self.__printDebug( "Error: Unable to bind to port [%s] - client will not be registered" % self.client_update_port, 0)
-            return    
+            self.__printDebug(
+                "Error: Unable to bind to port [%s] - client will not be registered" % self.client_update_port, 0)
+            return
 
         update_sock.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_TTL, 255)
-        status = update_sock.setsockopt(socket.IPPROTO_IP, socket.IP_ADD_MEMBERSHIP, socket.inet_aton(self._multicast_address) + socket.inet_aton('0.0.0.0'))
+        status = update_sock.setsockopt(socket.IPPROTO_IP, socket.IP_ADD_MEMBERSHIP,
+                                        socket.inet_aton(self._multicast_address) + socket.inet_aton('0.0.0.0'))
         update_sock.setblocking(0)
         self.__printDebug("Sending registration data: HELLO %s\n%s" % (self.client_header, self.client_data), 3)
 
-        #Send initial client registration
+        # Send initial client registration
         try:
             update_sock.sendto("HELLO %s\n%s" % (self.client_header, self.client_data), self.client_register_group)
         except:
-            self.__printDebug( "Error: Unable to send registration message" , 0)
+            self.__printDebug("Error: Unable to send registration message", 0)
 
-        #Now, listen for client discovery requests and respond.
+        # Now, listen for client discovery requests and respond.
         while self._registration_is_running:
             try:
                 data, addr = update_sock.recvfrom(1024)
@@ -102,24 +105,24 @@ class plexgdm:
                 pass
             else:
                 if "M-SEARCH * HTTP/1." in data:
-                    self.__printDebug("Detected client discovery request from %s.  Replying" % ( addr ,) , 2)
+                    self.__printDebug("Detected client discovery request from %s.  Replying" % (addr,), 2)
                     try:
                         update_sock.sendto("HTTP/1.0 200 OK\n%s" % self.client_data, addr)
                     except:
-                        self.__printDebug( "Error: Unable to send client update message",0)
+                        self.__printDebug("Error: Unable to send client update message", 0)
 
                     self.__printDebug("Sending registration data: HTTP/1.0 200 OK\n%s" % (self.client_data), 3)
                     self.client_registered = True
-            time.sleep(0.5)        
+            time.sleep(0.5)
 
-        self.__printDebug("Client Update loop stopped",1)
+        self.__printDebug("Client Update loop stopped", 1)
 
-        #When we are finished, then send a final goodbye message to de-register cleanly.
+        # When we are finished, then send a final goodbye message to de-register cleanly.
         self.__printDebug("Sending registration data: BYE %s\n%s" % (self.client_header, self.client_data), 3)
         try:
             update_sock.sendto("BYE %s\n%s" % (self.client_header, self.client_data), self.client_register_group)
         except:
-            self.__printDebug( "Error: Unable to send client update message" ,0)
+            self.__printDebug("Error: Unable to send client update message", 0)
 
         self.client_registered = False
 
@@ -128,22 +131,22 @@ class plexgdm:
         if self.client_registered and self.discovery_complete:
 
             if not self.server_list:
-                self.__printDebug("Server list is empty. Unable to check",2)
+                self.__printDebug("Server list is empty. Unable to check", 2)
                 return False
 
             try:
-                media_server=self.server_list[0]['server']
-                media_port=self.server_list[0]['port']
+                media_server = self.server_list[0]['server']
+                media_port = self.server_list[0]['port']
 
-                self.__printDebug("Checking server [%s] on port [%s]" % (media_server, media_port) ,2)                    
+                self.__printDebug("Checking server [%s] on port [%s]" % (media_server, media_port), 2)
                 f = urllib2.urlopen('http://%s:%s/clients' % (media_server, media_port))
                 client_result = f.read()
                 if self.client_id in client_result:
-                    self.__printDebug("Client registration successful",1)
+                    self.__printDebug("Client registration successful", 1)
                     self.__printDebug("Client data is: %s" % client_result, 3)
                     return True
                 else:
-                    self.__printDebug("Client registration not found",1)
+                    self.__printDebug("Client registration not found", 1)
                     self.__printDebug("Client data is: %s" % client_result, 3)
 
             except:
@@ -152,7 +155,7 @@ class plexgdm:
 
         return False
 
-    def getServerList (self):
+    def getServerList(self):
         return self.server_list
 
     def discover(self):
@@ -177,8 +180,8 @@ class plexgdm:
                     data, server = sock.recvfrom(1024)
                     self.__printDebug("Received data from %s, %s" % server, 3)
                     self.__printDebug("Data received is:\n %s" % data, 3)
-                    returnData.append( { 'from' : server,
-                                         'data' : data } )
+                    returnData.append({'from': server,
+                                       'data': data})
                 except socket.timeout:
                     break
         finally:
@@ -191,16 +194,16 @@ class plexgdm:
         if returnData:
 
             for response in returnData:
-                update = { 'server' : response.get('from')[0] }
+                update = {'server': response.get('from')[0]}
 
-                #Check if we had a positive HTTP response                        
+                # Check if we had a positive HTTP response
                 if "200 OK" in response.get('data'):
 
                     update['discovery'] = "auto"
-                    update['owned']='1'
-                    update['master']= 1
-                    update['role']='master'
-                    update['class']=None
+                    update['owned'] = '1'
+                    update['master'] = 1
+                    update['role'] = 'master'
+                    update['class'] = None
 
                     for each in response.get('data').split('\n'):
 
@@ -219,16 +222,16 @@ class plexgdm:
                         elif "Server-Class:" in each:
                             update['class'] = each.split(':')[1].strip()
 
-                discovered_servers.append(update)                    
+                discovered_servers.append(update)
 
         self.server_list = discovered_servers
 
         if not self.server_list:
-            self.__printDebug("No servers have been discovered",1)
+            self.__printDebug("No servers have been discovered", 1)
         else:
-            self.__printDebug("Number of servers Discovered: %s" % len(self.server_list),1)
+            self.__printDebug("Number of servers Discovered: %s" % len(self.server_list), 1)
             for items in self.server_list:
-                self.__printDebug("Server Discovered: %s" % items['serverName'] ,2)
+                self.__printDebug("Server Discovered: %s" % items['serverName'], 2)
 
     def setInterval(self, interval):
         self.discovery_interval = interval
@@ -256,18 +259,18 @@ class plexgdm:
             self.__printDebug("Registration not running", 1)
 
     def run_discovery_loop(self):
-        #Run initial discovery
+        # Run initial discovery
         self.discover()
 
-        discovery_count=0
+        discovery_count = 0
         while self._discovery_is_running:
-            discovery_count+=1
+            discovery_count += 1
             if discovery_count > self.discovery_interval:
                 self.discover()
-                discovery_count=0
+                discovery_count = 0
             time.sleep(1)
 
-    def start_discovery(self, daemon = False):
+    def start_discovery(self, daemon=False):
         if not self._discovery_is_running:
             self.__printDebug("Discovery starting up", 1)
             self._discovery_is_running = True
@@ -277,7 +280,7 @@ class plexgdm:
         else:
             self.__printDebug("Discovery already running", 1)
 
-    def start_registration(self, daemon = False):
+    def start_registration(self, daemon=False):
         if not self._registration_is_running:
             self.__printDebug("Registration starting up", 1)
             self._registration_is_running = True
@@ -287,11 +290,11 @@ class plexgdm:
         else:
             self.__printDebug("Registration already running", 1)
 
-    def start_all(self, daemon = False):
+    def start_all(self, daemon=False):
         self.start_discovery(daemon)
         self.start_registration(daemon)
 
-#Example usage
+# Example usage
 if __name__ == '__main__':
     client = plexgdm(debug=3)
     client.clientDetails("Test-Name", "Test Client", "3003", "Test-App", "1.2.3")
